refactor(dashboard): split 4 monolithic stores into signals/actions/normalizers
Apply the mission-store pattern (barrel re-export + separate files) to:
- governance-store (148 → signals + actions)
- room-truth-store (216 → signals + normalizers + actions)
- command-store (406 → signals + actions, normalizers already separate)
- operator-store (581 → signals + normalizers + actions)

Barrel re-exports preserve all existing import paths.
Skipped proof-store (22 lines), live-store and observatory-store
(pure computed signals with no async actions — splitting adds no value).
