+++ dashboard/src/operator-store.ts
@@ -1,581 +1,24 @@
-import { signal } from '@preact/signals'
-import { confirmOperatorAction, fetchOperatorDigest, fetchOperatorSnapshot, runOperatorAction } from './api'
-import type {
-  Message,
-  OperatorActionLogEntry,
-  OperatorActionDescriptor,
-  OperatorActionRequest,
-  OperatorActionResult,
-  OperatorAttentionItem,
-  OperatorDigest,
-  OperatorGuidanceSummary,
-  OperatorJudgment,
-  OperatorKeeperSnapshot,
-  OperatorLinkedAutoresearch,
-  OperatorResidentJudgeRuntime,
-  OperatorRecommendedAction,
-  OperatorSessionSnapshot,
-  OperatorSessionCard,
-  OperatorSnapshot,
-  OperatorRoomSnapshot,
-  PendingConfirmation,
-  OperatorWorkerCard,
-} from './types'
-import { registerOperatorRefresh } from './sse-store'
-import { isRecord, asString, asNumber, asBoolean, asStringArray, extractArray } from './components/common/normalize'
-import {
-  normalizeOperatorActionDescriptor,
-  normalizePendingConfirmation,
-  normalizePendingConfirmEnvelope,
-  normalizePendingConfirmSummary,
-} from './pending-confirm'
-
-export const operatorSnapshot = signal<OperatorSnapshot | null>(null)
-export const operatorRoomDigest = signal<OperatorDigest | null>(null)
-export const operatorSessionDigest = signal<OperatorDigest | null>(null)
-export const operatorLoading = signal(false)
-export const operatorError = signal<string | null>(null)
-export const operatorDigestLoading = signal(false)
-export const operatorDigestError = signal<string | null>(null)
-export const operatorActionBusy = signal(false)
-export const operatorActionLog = signal<OperatorActionLogEntry[]>([])
-
-let nextLogId = 1
-
-interface RefreshOptions {
-  force?: boolean
-}
-
-const OPERATOR_REFRESH_TTL_MS = 1_000
-
-let snapshotRefreshInflight: Promise<void> | null = null
-let roomDigestRefreshInflight: Promise<void> | null = null
-let sessionDigestRefreshInflight: Promise<void> | null = null
-let lastSnapshotRefreshAt = 0
-let lastRoomDigestRefreshAt = 0
-let lastSessionDigestRefreshAt = 0
-let lastSessionDigestId: string | null = null
-
-function normalizeMessage(raw: unknown): Message | null {
-  if (!isRecord(raw)) return null
-  return {
-    id: asString(raw.id),
-    seq: asNumber(raw.seq),
-    from: asString(raw.from) ?? asString(raw.from_agent) ?? 'system',
-    content: asString(raw.content) ?? '',
-    timestamp: asString(raw.timestamp) ?? new Date().toISOString(),
-    type: asString(raw.type),
-  }
-}
-
-function normalizeRoom(raw: unknown): OperatorRoomSnapshot {
-  if (!isRecord(raw)) return {}
-  return {
-    room_id: asString(raw.room_id),
-    current_room: asString(raw.current_room) ?? asString(raw.room),
-    project: asString(raw.project),
-    cluster: asString(raw.cluster),
-    paused: asBoolean(raw.paused),
-    pause_reason: asString(raw.pause_reason) ?? null,
-    paused_by: asString(raw.paused_by) ?? null,
-    paused_at: asString(raw.paused_at) ?? null,
-  }
-}
-
-function normalizeStringRecord(raw: unknown): Record<string, string> | undefined {
-  if (!isRecord(raw)) return undefined
-  const entries = Object.entries(raw)
-    .map(([key, value]) => {
-      const text = asString(value)
-      return text ? [key, text] : null
-    })
-    .filter((entry): entry is [string, string] => entry !== null)
-  return entries.length > 0 ? Object.fromEntries(entries) : undefined
-}
-
-function normalizeAttentionItem(raw: unknown): OperatorAttentionItem | null {
-  if (!isRecord(raw)) return null
-  const kind = asString(raw.kind)
-  const summary = asString(raw.summary)
-  const targetType = asString(raw.target_type)
-  if (!kind || !summary || !targetType) return null
-  return {
-    kind,
-    severity: asString(raw.severity) ?? 'unknown',
-    summary,
-    target_type: targetType,
-    target_id: asString(raw.target_id) ?? null,
-    actor: asString(raw.actor) ?? null,
-    evidence: raw.evidence,
-  }
-}
-
-function normalizeRecommendedAction(raw: unknown): OperatorRecommendedAction | null {
-  if (!isRecord(raw)) return null
-  const actionType = asString(raw.action_type)
-  const targetType = asString(raw.target_type)
-  const reason = asString(raw.reason)
-  if (!actionType || !targetType || !reason) return null
-  return {
-    action_type: actionType,
-    target_type: targetType,
-    target_id: asString(raw.target_id) ?? null,
-    severity: asString(raw.severity) ?? 'unknown',
-    reason,
-    confirm_required: asBoolean(raw.confirm_required),
-    suggested_payload: raw.suggested_payload,
-    preview: raw.preview,
-  }
-}
-
-function normalizeResidentJudgeRuntime(raw: unknown): OperatorResidentJudgeRuntime | null {
-  if (!isRecord(raw)) return null
-  return {
-    enabled: asBoolean(raw.enabled),
-    judge_online: asBoolean(raw.judge_online),
-    refreshing: asBoolean(raw.refreshing),
-    generated_at: asString(raw.generated_at) ?? null,
-    expires_at: asString(raw.expires_at) ?? null,
-    model_used: asString(raw.model_used) ?? null,
-    keeper_name: asString(raw.keeper_name) ?? null,
-    last_error: asString(raw.last_error) ?? null,
-  }
-}
-
-function normalizeGuidanceSummary(raw: unknown): OperatorGuidanceSummary | null {
-  if (!isRecord(raw)) return null
-  return {
-    summary: asString(raw.summary) ?? null,
-    confidence: asNumber(raw.confidence) ?? null,
-    provenance: asString(raw.provenance) ?? null,
-    authoritative: asBoolean(raw.authoritative),
-    surface: asString(raw.surface) ?? null,
-    fresh_until: asString(raw.fresh_until) ?? null,
-    keeper_name: asString(raw.keeper_name) ?? null,
-    fallback_used: asBoolean(raw.fallback_used),
-    disagreement_with_truth: asBoolean(raw.disagreement_with_truth),
-  }
-}
-
-function normalizeOperatorJudgment(raw: unknown): OperatorJudgment | null {
-  if (!isRecord(raw)) return null
-  return {
-    judgment_id: asString(raw.judgment_id) ?? undefined,
-    surface: asString(raw.surface) ?? null,
-    target_type: asString(raw.target_type) ?? null,
-    target_id: asString(raw.target_id) ?? null,
-    status: asString(raw.status) ?? null,
-    summary: asString(raw.summary) ?? null,
-    confidence: asNumber(raw.confidence) ?? null,
-    generated_at: asString(raw.generated_at) ?? null,
-    fresh_until: asString(raw.fresh_until) ?? null,
-    keeper_name: asString(raw.keeper_name) ?? null,
-    model_name: asString(raw.model_name) ?? null,
-    runtime_name: asString(raw.runtime_name) ?? null,
-    evidence_refs: asStringArray(raw.evidence_refs),
-    recommended_action: normalizeRecommendedAction(raw.recommended_action),
-    supersedes: asStringArray(raw.supersedes),
-    fallback_used: asBoolean(raw.fallback_used),
-    disagreement_with_truth: asBoolean(raw.disagreement_with_truth),
-    provenance: asString(raw.provenance) ?? null,
-  }
-}
-
-function normalizeWorkerCard(raw: unknown): OperatorWorkerCard | null {
-  if (!isRecord(raw)) return null
-  return {
-    actor: asString(raw.actor) ?? null,
-    spawn_agent: asString(raw.spawn_agent) ?? null,
-    spawn_role: asString(raw.spawn_role) ?? null,
-    spawn_model: asString(raw.spawn_model) ?? null,
-    worker_class: asString(raw.worker_class) ?? null,
-    parent_actor: asString(raw.parent_actor) ?? null,
-    capsule_mode: asString(raw.capsule_mode) ?? null,
-    runtime_pool: asString(raw.runtime_pool) ?? null,
-    lane_id: asString(raw.lane_id) ?? null,
-    controller_level: asString(raw.controller_level) ?? null,
-    control_domain: asString(raw.control_domain) ?? null,
-    supervisor_actor: asString(raw.supervisor_actor) ?? null,
-    model_tier: asString(raw.model_tier) ?? null,
-    task_profile: asString(raw.task_profile) ?? null,
-    risk_level: asString(raw.risk_level) ?? null,
-    routing_confidence: asNumber(raw.routing_confidence) ?? null,
-    routing_reason: asString(raw.routing_reason) ?? null,
-    status: asString(raw.status) ?? 'unknown',
-    turn_count: asNumber(raw.turn_count) ?? 0,
-    empty_note_turn_count: asNumber(raw.empty_note_turn_count) ?? 0,
-    has_turn: asBoolean(raw.has_turn) ?? false,
-    last_turn_ts_iso: asString(raw.last_turn_ts_iso) ?? null,
-  }
-}
-
-function normalizeSessionCard(raw: unknown): OperatorSessionCard | null {
-  if (!isRecord(raw)) return null
-  const sessionId = asString(raw.session_id)
-  if (!sessionId) return null
-  return {
-    session_id: sessionId,
-    goal: asString(raw.goal),
-    status: asString(raw.status),
-    health: asString(raw.health),
-    scale_profile: asString(raw.scale_profile),
-    control_profile: asString(raw.control_profile),
-    planned_worker_count: asNumber(raw.planned_worker_count),
-    active_agent_count: asNumber(raw.active_agent_count),
-    last_turn_age_sec: asNumber(raw.last_turn_age_sec) ?? null,
-    attention_count: asNumber(raw.attention_count),
-    recommended_action_count: asNumber(raw.recommended_action_count),
-    top_attention: normalizeAttentionItem(raw.top_attention),
-    top_recommendation: normalizeRecommendedAction(raw.top_recommendation),
-  }
-}
-
-function normalizeLinkedAutoresearch(raw: unknown): OperatorLinkedAutoresearch | null {
-  if (!isRecord(raw)) return null
-  const loopId = asString(raw.loop_id)
-  const status = asString(raw.status)
-  if (!loopId && !status) return null
-  return {
-    loop_id: loopId ?? null,
-    session_id: asString(raw.session_id) ?? null,
-    status: status ?? null,
-    current_cycle: asNumber(raw.current_cycle) ?? undefined,
-    best_score: asNumber(raw.best_score) ?? null,
-    last_decision: asString(raw.last_decision) ?? null,
-    target_file: asString(raw.target_file) ?? null,
-    workdir: asString(raw.workdir) ?? null,
-    source_workdir: asString(raw.source_workdir) ?? null,
-    program_note: asString(raw.program_note) ?? null,
-    operation_id: asString(raw.operation_id) ?? null,
-    queued_hypothesis: asString(raw.queued_hypothesis) ?? null,
-    warnings: extractArray(raw.warnings)
-      .map(item => (typeof item === 'string' ? item.trim() : ''))
-      .filter(Boolean),
-    error: asString(raw.error) ?? null,
-  }
-}
-
-function normalizeOperatorDigest(raw: unknown): OperatorDigest {
-  const root = isRecord(raw) ? raw : {}
-  return {
-    trace_id: asString(root.trace_id),
-    target_type: asString(root.target_type) ?? 'room',
-    target_id: asString(root.target_id) ?? null,
-    health: asString(root.health),
-    judgment_owner: asString(root.judgment_owner) ?? null,
-    authoritative_judgment_available: asBoolean(root.authoritative_judgment_available),
-    resident_judge_runtime: normalizeResidentJudgeRuntime(root.resident_judge_runtime),
-    judgment: normalizeOperatorJudgment(root.judgment),
-    active_guidance_layer: asString(root.active_guidance_layer) ?? null,
-    active_summary: normalizeGuidanceSummary(root.active_summary),
-    active_recommended_actions: extractArray(root.active_recommended_actions)
-      .map(normalizeRecommendedAction)
-      .filter((item): item is OperatorRecommendedAction => item !== null),
-    active_recommendation_source: asString(root.active_recommendation_source) ?? null,
-    active_recommendation_summary: normalizeGuidanceSummary(root.active_recommendation_summary),
-    fallback_recommended_actions: extractArray(root.fallback_recommended_actions)
-      .map(normalizeRecommendedAction)
-      .filter((item): item is OperatorRecommendedAction => item !== null),
-    recommendation_summary: normalizeGuidanceSummary(root.recommendation_summary),
-    swarm_status: isRecord(root.swarm_status)
-      ? (root.swarm_status as unknown as OperatorDigest['swarm_status'])
-      : undefined,
-    attention_items: extractArray(root.attention_items)
-      .map(normalizeAttentionItem)
-      .filter((item): item is OperatorAttentionItem => item !== null),
-    recommended_actions: extractArray(root.recommended_actions)
-      .map(normalizeRecommendedAction)
-      .filter((item): item is OperatorRecommendedAction => item !== null),
-    session_cards: extractArray(root.session_cards)
-      .map(normalizeSessionCard)
-      .filter((item): item is OperatorSessionCard => item !== null),
-    worker_cards: extractArray(root.worker_cards)
-      .map(normalizeWorkerCard)
-      .filter((item): item is OperatorWorkerCard => item !== null),
-  }
-}
-
-function normalizeSession(raw: unknown): OperatorSessionSnapshot | null {
-  if (!isRecord(raw)) return null
-  const statusBlock = isRecord(raw.status) ? raw.status : undefined
-  const summary = isRecord(raw.summary) ? raw.summary : isRecord(statusBlock?.summary) ? statusBlock.summary : undefined
-  const session = isRecord(raw.session) ? raw.session : isRecord(statusBlock?.session) ? statusBlock.session : undefined
-  const sessionId =
-    asString(raw.session_id)
-    ?? asString(summary?.session_id)
-    ?? asString(session?.session_id)
-  if (!sessionId) return null
-
-  const reportPaths = normalizeStringRecord(raw.report_paths)
-    ?? normalizeStringRecord(statusBlock?.report_paths)
-  const recentEvents = extractArray(raw.recent_events, ['events'])
-    .filter(isRecord)
-
-  return {
-    session_id: sessionId,
-    status: asString(raw.status) ?? asString(summary?.status) ?? asString(session?.status),
-    progress_pct: asNumber(raw.progress_pct) ?? asNumber(summary?.progress_pct),
-    elapsed_sec: asNumber(raw.elapsed_sec) ?? asNumber(summary?.elapsed_sec),
-    remaining_sec: asNumber(raw.remaining_sec) ?? asNumber(summary?.remaining_sec),
-    done_delta_total: asNumber(raw.done_delta_total) ?? asNumber(summary?.done_delta_total),
-    summary,
-    team_health: isRecord(raw.team_health) ? raw.team_health : isRecord(statusBlock?.team_health) ? statusBlock.team_health : undefined,
-    communication_metrics: isRecord(raw.communication_metrics) ? raw.communication_metrics : isRecord(statusBlock?.communication_metrics) ? statusBlock.communication_metrics : undefined,
-    orchestration_state: isRecord(raw.orchestration_state) ? raw.orchestration_state : isRecord(statusBlock?.orchestration_state) ? statusBlock.orchestration_state : undefined,
-    cascade_metrics: isRecord(raw.cascade_metrics) ? raw.cascade_metrics : isRecord(statusBlock?.cascade_metrics) ? statusBlock.cascade_metrics : undefined,
-    report_paths: reportPaths,
-    linked_autoresearch:
-      normalizeLinkedAutoresearch(raw.linked_autoresearch)
-      ?? normalizeLinkedAutoresearch(statusBlock?.linked_autoresearch)
-      ?? null,
-    session,
-    recent_events: recentEvents,
-  }
-}
-
-function normalizeKeeper(raw: unknown): OperatorKeeperSnapshot | null {
-  if (!isRecord(raw)) return null
-  const name = asString(raw.name)
-  if (!name) return null
-  const contextRaw = isRecord(raw.context) ? raw.context : undefined
-  return {
-    name,
-    runtime_class:
-      raw.runtime_class === 'persistent_agent' ? 'persistent_agent' : 'resident_keeper',
-    desired: asBoolean(raw.desired),
-    resident_registered: asBoolean(raw.resident_registered),
-    agent_name: asString(raw.agent_name),
-    status: asString(raw.status),
-    context_ratio: asNumber(raw.context_ratio) ?? asNumber(contextRaw?.context_ratio),
-    generation: asNumber(raw.generation),
-    active_goal_ids: asStringArray(raw.active_goal_ids),
-    last_autonomous_action_at: asString(raw.last_autonomous_action_at) ?? null,
-    last_turn_ago_s: asNumber(raw.last_turn_ago_s),
-    model: asString(raw.model) ?? asString(raw.active_model) ?? asString(raw.primary_model),
-  }
-}
-
-function normalizeOperatorSnapshot(raw: unknown): OperatorSnapshot {
-  const root = isRecord(raw) ? raw : {}
-  const pendingConfirmEnvelope = normalizePendingConfirmEnvelope(root.pending_confirm_envelope)
-  return {
-    room: normalizeRoom(root.room),
-    sessions: extractArray(root.sessions, ['items', 'sessions'])
-      .map(normalizeSession)
-      .filter((item): item is OperatorSessionSnapshot => item !== null),
-    keepers: extractArray(root.keepers, ['items', 'keepers'])
-      .map(normalizeKeeper)
-      .filter((item): item is OperatorKeeperSnapshot => item !== null),
-    resident_judge_runtime: normalizeResidentJudgeRuntime(root.resident_judge_runtime),
-    persistent_agents: extractArray(root.persistent_agents, ['items', 'persistent_agents'])
-      .map(normalizeKeeper)
-      .filter((item): item is OperatorKeeperSnapshot => item !== null),
-    recent_messages: extractArray(root.recent_messages, ['messages'])
-      .map(normalizeMessage)
-      .filter((item): item is Message => item !== null),
-    pending_confirms: pendingConfirmEnvelope?.items
-      ?? extractArray(root.pending_confirms, ['items', 'confirms'])
-        .map(normalizePendingConfirmation)
-        .filter((item): item is PendingConfirmation => item !== null),
-    pending_confirm_envelope: pendingConfirmEnvelope ?? undefined,
-    pending_confirm_summary:
-      pendingConfirmEnvelope?.summary
-      ?? normalizePendingConfirmSummary(root.pending_confirm_summary)
-      ?? undefined,
-    available_actions: extractArray(root.available_actions, ['actions'])
-      .map(normalizeOperatorActionDescriptor)
-      .filter((item): item is OperatorActionDescriptor => item !== null),
-  }
-}
-
-function stringifyUnknown(value: unknown): string {
-  if (typeof value === 'string') return value
-  if (value === null || value === undefined) return ''
-  try {
-    return JSON.stringify(value)
-  } catch {
-    return String(value)
-  }
-}
-
-function targetLabelOf(request: Pick<OperatorActionRequest, 'target_type' | 'target_id'>): string {
-  return request.target_id ? `${request.target_type}:${request.target_id}` : request.target_type
-}
-
-function appendLog(entry: Omit<OperatorActionLogEntry, 'id' | 'at'>): void {
-  operatorActionLog.value = [
-    {
-      ...entry,
-      id: nextLogId++,
-      at: new Date().toISOString(),
-    },
-    ...operatorActionLog.value,
-  ].slice(0, 20)
-}
-
-function logMessageFromResult(result: OperatorActionResult): string {
-  if (result.confirm_required) {
-    return stringifyUnknown(result.preview) || 'Confirmation required'
-  }
-  return stringifyUnknown(result.result)
-    || stringifyUnknown(result.executed_action)
-    || stringifyUnknown(result.delegated_tool_result)
-    || result.status
-}
-
-function isFresh(lastAt: number, opts?: RefreshOptions): boolean {
-  return !opts?.force && Date.now() - lastAt < OPERATOR_REFRESH_TTL_MS
-}
-
-export async function refreshOperatorSnapshot(opts?: RefreshOptions): Promise<void> {
-  if (snapshotRefreshInflight) return snapshotRefreshInflight
-  if (isFresh(lastSnapshotRefreshAt, opts)) return
-  operatorLoading.value = true
-  operatorError.value = null
-  snapshotRefreshInflight = (async () => {
-    try {
-      const raw = await fetchOperatorSnapshot()
-      operatorSnapshot.value = normalizeOperatorSnapshot(raw)
-      lastSnapshotRefreshAt = Date.now()
-    } catch (err) {
-      operatorError.value = err instanceof Error ? err.message : 'Failed to load operator snapshot'
-    } finally {
-      operatorLoading.value = false
-      snapshotRefreshInflight = null
-    }
-  })()
-  return snapshotRefreshInflight
-}
-
-export async function refreshOperatorRoomDigest(opts?: RefreshOptions): Promise<void> {
-  if (roomDigestRefreshInflight) return roomDigestRefreshInflight
-  if (isFresh(lastRoomDigestRefreshAt, opts)) return
-  operatorDigestLoading.value = true
-  operatorDigestError.value = null
-  roomDigestRefreshInflight = (async () => {
-    try {
-      const raw = await fetchOperatorDigest({ targetType: 'room' })
-      operatorRoomDigest.value = normalizeOperatorDigest(raw)
-      lastRoomDigestRefreshAt = Date.now()
-    } catch (err) {
-      operatorDigestError.value = err instanceof Error ? err.message : 'Failed to load operator digest'
-    } finally {
-      operatorDigestLoading.value = false
-      roomDigestRefreshInflight = null
-    }
-  })()
-  return roomDigestRefreshInflight
-}
-
-export async function refreshOperatorSessionDigest(sessionId: string | null, opts?: RefreshOptions): Promise<void> {
-  if (!sessionId) {
-    operatorSessionDigest.value = null
-    lastSessionDigestId = null
-    return
-  }
-  if (sessionDigestRefreshInflight && lastSessionDigestId === sessionId) return sessionDigestRefreshInflight
-  if (lastSessionDigestId === sessionId && isFresh(lastSessionDigestRefreshAt, opts)) return
-
-  operatorDigestLoading.value = true
-  operatorDigestError.value = null
-  lastSessionDigestId = sessionId
-  sessionDigestRefreshInflight = (async () => {
-    try {
-      const raw = await fetchOperatorDigest({
-        targetType: 'team_session',
-        targetId: sessionId,
-        includeWorkers: true,
-      })
-      operatorSessionDigest.value = normalizeOperatorDigest(raw)
-      lastSessionDigestRefreshAt = Date.now()
-    } catch (err) {
-      operatorDigestError.value = err instanceof Error ? err.message : 'Failed to load session digest'
-    } finally {
-      operatorDigestLoading.value = false
-      sessionDigestRefreshInflight = null
-    }
-  })()
-  return sessionDigestRefreshInflight
-}
-
-export async function dispatchOperatorAction(request: OperatorActionRequest): Promise<OperatorActionResult> {
-  operatorActionBusy.value = true
-  operatorError.value = null
-  try {
-    const result = await runOperatorAction(request)
-    appendLog({
-      actor: request.actor,
-      action_type: request.action_type,
-      target_label: targetLabelOf(request),
-      outcome: result.confirm_required ? 'preview' : 'executed',
-      message: logMessageFromResult(result),
-      delegated_tool: result.delegated_tool,
-    })
-    await refreshOperatorSnapshot({ force: true })
-    await refreshOperatorRoomDigest({ force: true })
-    if (operatorSessionDigest.value?.target_id) {
-      await refreshOperatorSessionDigest(operatorSessionDigest.value.target_id, { force: true })
-    }
-    return result
-  } catch (err) {
-    const message = err instanceof Error ? err.message : 'Operator action failed'
-    operatorError.value = message
-    appendLog({
-      actor: request.actor,
-      action_type: request.action_type,
-      target_label: targetLabelOf(request),
-      outcome: 'error',
-      message,
-    })
-    throw err
-  } finally {
-    operatorActionBusy.value = false
-  }
-}
-
-export async function confirmOperatorPendingAction(
-  actor: string,
-  confirmToken: string,
-  decision: 'confirm' | 'deny' = 'confirm',
-): Promise<OperatorActionResult> {
-  operatorActionBusy.value = true
-  operatorError.value = null
-  try {
-    const result = await confirmOperatorAction(actor, confirmToken, decision)
-    appendLog({
-      actor,
-      action_type: decision,
-      target_label: confirmToken,
-      outcome: 'confirmed',
-      message: logMessageFromResult(result),
-      delegated_tool: result.delegated_tool,
-    })
-    await refreshOperatorSnapshot({ force: true })
-    await refreshOperatorRoomDigest({ force: true })
-    if (operatorSessionDigest.value?.target_id) {
-      await refreshOperatorSessionDigest(operatorSessionDigest.value.target_id, { force: true })
-    }
-    return result
-  } catch (err) {
-    const message = err instanceof Error ? err.message : 'Operator confirmation failed'
-    operatorError.value = message
-    appendLog({
-      actor,
-      action_type: 'confirm',
-      target_label: confirmToken,
-      outcome: 'error',
-      message,
-    })
-    throw err
-  } finally {
-    operatorActionBusy.value = false
-  }
-}
-
-registerOperatorRefresh(() => {
-  void refreshOperatorSnapshot({ force: true })
-  void refreshOperatorRoomDigest({ force: true })
-  if (operatorSessionDigest.value?.target_id) {
-    void refreshOperatorSessionDigest(operatorSessionDigest.value.target_id, { force: true })
-  }
-})
+// Barrel re-export for backward compatibility.
+// Operator store is split into operator-signals, operator-normalizers, operator-actions.
+export {
+  operatorSnapshot,
+  operatorRoomDigest,
+  operatorSessionDigest,
+  operatorLoading,
+  operatorError,
+  operatorDigestLoading,
+  operatorDigestError,
+  operatorActionBusy,
+  operatorActionLog,
+} from './operator-signals'
+export {
+  normalizeOperatorDigest,
+  normalizeOperatorSnapshot,
+} from './operator-normalizers'
+export {
+  refreshOperatorSnapshot,
+  refreshOperatorRoomDigest,
+  refreshOperatorSessionDigest,
+  dispatchOperatorAction,
+  confirmOperatorPendingAction,
+} from './operator-actions'
