+++ dashboard/src/room-truth-store.ts
@@ -1,216 +1,10 @@
-import { signal } from '@preact/signals'
-import { fetchDashboardRoomTruth } from './api'
-import type {
-  DashboardRoomTruthAttentionSummary,
-  DashboardRoomTruthFocus,
-  DashboardRoomTruthRecommendationSummary,
-  DashboardRoomTruthResponse,
-  PendingConfirmSummary,
-  ServerStatus,
-} from './types'
-import { asBoolean, asNumber, asString, asStringArray, isRecord, extractArray } from './components/common/normalize'
-import {
-  normalizeExecutionSummary,
-  normalizeExecutionQueueItem,
-  normalizeAttentionItem,
-  normalizeRecommendedAction,
-} from './store-normalizers'
-
-export const roomTruth = signal<DashboardRoomTruthResponse | null>(null)
-export const roomTruthLoading = signal(false)
-export const roomTruthError = signal<string | null>(null)
-
-let inflightRoomTruthRefresh: Promise<void> | null = null
-let lastRoomTruthRefreshAt = 0
-const ROOM_TRUTH_TTL_MS = 60_000
-
-function normalizeServerStatus(raw: unknown): ServerStatus | null {
-  if (!isRecord(raw)) return null
-  return {
-    room: asString(raw.room) ?? asString(raw.current_room),
-    room_base_path: asString(raw.room_base_path),
-    cluster: asString(raw.cluster),
-    project: asString(raw.project),
-    paused: asBoolean(raw.paused),
-    version: asString(raw.version),
-    generated_at: asString(raw.generated_at),
-    tempo_interval_s: asNumber(raw.tempo_interval_s),
-  }
-}
-
-function normalizePendingConfirmSummary(raw: unknown): PendingConfirmSummary | null {
-  if (!isRecord(raw)) return null
-  return {
-    actor_filter: asString(raw.actor_filter) ?? null,
-    filter_active: asBoolean(raw.filter_active) ?? false,
-    visible_count: asNumber(raw.visible_count) ?? 0,
-    total_count: asNumber(raw.total_count) ?? 0,
-    hidden_count: asNumber(raw.hidden_count) ?? 0,
-    hidden_actors: asStringArray(raw.hidden_actors),
-    confirm_required_actions: extractArray(raw.confirm_required_actions).flatMap(item => {
-      if (!isRecord(item)) return []
-      const actionType = asString(item.action_type)
-      const targetType = asString(item.target_type)
-      if (!actionType || !targetType) return []
-      return [{
-        action_type: actionType,
-        target_type: targetType,
-        description: asString(item.description),
-        confirm_required: asBoolean(item.confirm_required),
-      }]
-    }),
-  }
-}
-
-function normalizeAttentionSummary(raw: unknown): DashboardRoomTruthAttentionSummary | null {
-  if (!isRecord(raw)) return null
-  return {
-    count: asNumber(raw.count) ?? 0,
-    bad_count: asNumber(raw.bad_count) ?? 0,
-    warn_count: asNumber(raw.warn_count) ?? 0,
-    provenance: asString(raw.provenance) ?? null,
-    top_item: normalizeAttentionItem(raw.top_item),
-  }
-}
-
-function normalizeRecommendationSummary(raw: unknown): DashboardRoomTruthRecommendationSummary | null {
-  if (!isRecord(raw)) return null
-  return {
-    count: asNumber(raw.count) ?? 0,
-    provenance: asString(raw.provenance) ?? null,
-    top_action: normalizeRecommendedAction(raw.top_action),
-  }
-}
-
-function normalizeFocus(raw: unknown): DashboardRoomTruthFocus | null {
-  if (!isRecord(raw)) return null
-  const label = asString(raw.label)
-  const reason = asString(raw.reason)
-  const source = asString(raw.source)
-  const provenance = asString(raw.provenance)
-  if (!label || !reason || !source || !provenance) return null
-  return {
-    label,
-    reason,
-    source,
-    provenance,
-    target_kind: asString(raw.target_kind) ?? null,
-    target_id: asString(raw.target_id) ?? null,
-    suggested_tab: asString(raw.suggested_tab) ?? null,
-    suggested_surface: asString(raw.suggested_surface) ?? null,
-    suggested_params: isRecord(raw.suggested_params)
-      ? Object.fromEntries(
-          Object.entries(raw.suggested_params)
-            .map(([key, value]) => {
-              const text = asString(value)
-              return text ? [key, text] : null
-            })
-            .filter((entry): entry is [string, string] => entry !== null),
-        )
-      : {},
-  }
-}
-
-function normalizeRoomTruth(raw: unknown): DashboardRoomTruthResponse {
-  const root = isRecord(raw) ? raw : {}
-  const roomBlock = isRecord(root.room) ? root.room : {}
-  const executionBlock = isRecord(root.execution) ? root.execution : {}
-  const commandBlock = isRecord(root.command) ? root.command : {}
-  const operatorBlock = isRecord(root.operator) ? root.operator : {}
-  return {
-    generated_at: asString(root.generated_at),
-    room: {
-      status: normalizeServerStatus(roomBlock.status),
-      counts: isRecord(roomBlock.counts)
-        ? {
-            agents: asNumber(roomBlock.counts.agents),
-            tasks: asNumber(roomBlock.counts.tasks),
-            keepers: asNumber(roomBlock.counts.keepers),
-          }
-        : undefined,
-      provenance: asString(roomBlock.provenance) ?? null,
-    },
-    execution: {
-      summary: normalizeExecutionSummary(executionBlock.summary),
-      top_queue: normalizeExecutionQueueItem(executionBlock.top_queue),
-      provenance: asString(executionBlock.provenance) ?? null,
-    },
-    command: {
-      active_operations: asNumber(commandBlock.active_operations),
-      active_detachments: asNumber(commandBlock.active_detachments),
-      pending_approvals: asNumber(commandBlock.pending_approvals),
-      bad_alerts: asNumber(commandBlock.bad_alerts),
-      warn_alerts: asNumber(commandBlock.warn_alerts),
-      moving_lanes: asNumber(commandBlock.moving_lanes),
-      active_lanes: asNumber(commandBlock.active_lanes),
-      provenance: asString(commandBlock.provenance) ?? null,
-    },
-    operator: {
-      health: asString(operatorBlock.health) ?? null,
-      attention_summary: normalizeAttentionSummary(operatorBlock.attention_summary),
-      recommendation_summary: normalizeRecommendationSummary(operatorBlock.recommendation_summary),
-      pending_confirm_summary: normalizePendingConfirmSummary(operatorBlock.pending_confirm_summary),
-      provenance: asString(operatorBlock.provenance) ?? null,
-    },
-    focus: normalizeFocus(root.focus),
-  }
-}
-
-/** True while the server is still computing its first cache (warm-up). */
-export const roomTruthInitializing = signal(false)
-
-const WARM_RETRY_DELAY_MS = 3_000
-const WARM_MAX_RETRIES = 10
-
-export async function refreshRoomTruth(opts?: { force?: boolean }): Promise<void> {
-  if (inflightRoomTruthRefresh) return inflightRoomTruthRefresh
-  if (!opts?.force && Date.now() - lastRoomTruthRefreshAt < ROOM_TRUTH_TTL_MS) return
-
-  roomTruthLoading.value = true
-  roomTruthError.value = null
-  inflightRoomTruthRefresh = (async () => {
-    try {
-      const raw = await fetchDashboardRoomTruth()
-      const isInitializing =
-        isRecord(raw)
-        && asString((raw as Record<string, unknown>).status) === 'initializing'
-      if (isInitializing) {
-        console.debug('[room-truth] server initializing, scheduling warm-up retry')
-        roomTruthInitializing.value = true
-        scheduleWarmRetry(1)
-        return
-      }
-      roomTruthInitializing.value = false
-      roomTruth.value = normalizeRoomTruth(raw)
-      lastRoomTruthRefreshAt = Date.now()
-    } catch (err) {
-      const detail = err instanceof Error ? err.message : 'Failed to load room truth'
-      console.warn('[room-truth] fetch failed:', detail)
-      roomTruthError.value = detail
-    } finally {
-      roomTruthLoading.value = false
-      inflightRoomTruthRefresh = null
-    }
-  })()
-
-  return inflightRoomTruthRefresh
-}
-
-function scheduleWarmRetry(attempt: number): void {
-  console.debug(`[room-truth] warm-up retry ${attempt}/${WARM_MAX_RETRIES}`)
-  if (attempt > WARM_MAX_RETRIES) {
-    roomTruthInitializing.value = false
-    roomTruthError.value = 'Server warm-up timed out. Try refreshing.'
-    roomTruthLoading.value = false
-    inflightRoomTruthRefresh = null
-    return
-  }
-  window.setTimeout(() => {
-    inflightRoomTruthRefresh = null
-    void refreshRoomTruth().then(() => {
-      if (roomTruthInitializing.value) {
-        scheduleWarmRetry(attempt + 1)
-      }
-    })
-  }, WARM_RETRY_DELAY_MS)
-}
+// Barrel re-export for backward compatibility.
+// Room truth store is split into room-truth-signals, room-truth-normalizers, room-truth-actions.
+export {
+  roomTruth,
+  roomTruthLoading,
+  roomTruthError,
+  roomTruthInitializing,
+} from './room-truth-signals'
+export { normalizeRoomTruth } from './room-truth-normalizers'
+export { refreshRoomTruth } from './room-truth-actions'
